requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
behave==1.2.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chromedriver-autoinstaller==0.6.2
cryptography==41.0.2
defusedxml==0.7.1
exceptiongroup==1.1.2
Faker==15.0.0
glob2==0.7
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
jaraco.classes==3.3.0
jira==3.1.1
keyring==24.2.0
Mako==1.2.4
MarkupSafe==2.1.3
more-itertools==10.0.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
parse==1.19.1
parse-type==0.6.2
pluggy==1.2.0
py==1.11.0
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
pytest==7.4.0
pytest-bdd==5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
selenium==4.1.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==1.26.16
urllib3-secure-extra==0.1.0
webdriver-manager==3.9.1
wsproto==1.2.0
zipp==3.16.2
